requirements.txt
streamlit
pdfplumber
spacy
textstat
google-generativeai
python-dotenv
numpy
pandas
